memory opt

+++ src/dtypes.py
@@ -5,7 +5,7 @@ def convert_to_dtypes(df, dtypes):
     for col in df.columns:
         t = dtypes[col]
         if df.dtypes[col].__str__() == "category":
-            df[col] = np.clip(df[col], np.iinfo(t).min, np.iinfo(t).max).astype(t)
+            df[col] = np.clip(df[col].astype(int), np.iinfo(t).min, np.iinfo(t).max).astype(t)
         elif t != df.dtypes[col]:
             if "float" in t.__str__():
                 df[col] = np.clip(df[col], np.finfo(t).min, np.finfo(t).max).astype(t)
